move cheatable syscall handler to new dir
