+++ crates/cheatnet/src/execution/calls.rs
@@ -1,3 +1,4 @@
+use crate::execution::syscalls::cairo1_cheatable_syscall_handler::CheatableSyscallHandler;
 use blockifier::{
     execution::{
         entry_point::{CallEntryPoint, CallType},
@@ -16,7 +17,7 @@ use starknet_api::{
     transaction::Calldata,
 };
 
-use super::{entry_point::execute_call_entry_point, syscalls::CheatableSyscallHandler};
+use super::{entry_point::execute_call_entry_point};
 
 // blockifier/src/execution/syscalls/hint_processor.rs:541 (execute_inner_call)
 pub fn execute_inner_call(
